feat: Update transaction handling to include money received; refactor related models and database schema

+++ lib/widgets/currency_input_formatter.dart
@@ -23,7 +23,7 @@ class CurrencyInputFormatter extends TextInputFormatter {
       return newValue;
     }
 
-    double value = double.parse(newValue.text.replaceAll(RegExp(r'[^0-9]'), ''));
+    int value = int.parse(newValue.text.replaceAll(RegExp(r'[^0-9]'), ''));
     String newText = _formatter.format(value);
 
     return newValue.copyWith(
